system: moved system helpers to separate namespace
